+++ packages/compiler-cli/src/ngtsc/entry_point/BUILD.bazel
@@ -1,0 +1,18 @@
+package(default_visibility = ["//visibility:public"])
+
+load("//tools:defaults.bzl", "ts_library")
+
+ts_library(
+    name = "entry_point",
+    srcs = glob([
+        "index.ts",
+        "src/**/*.ts",
+    ]),
+    module_name = "@angular/compiler-cli/src/ngtsc/entry_point",
+    deps = [
+        "//packages/compiler-cli/src/ngtsc/shims",
+        "//packages/compiler-cli/src/ngtsc/util",
+        "@ngdeps//@types/node",
+        "@ngdeps//typescript",
+    ],
+)
