+++ packages/compiler-cli/src/ngtsc/shims/index.ts
@@ -9,6 +9,5 @@
 /// <reference types="node" />
 
 export {FactoryGenerator, FactoryInfo, generatedFactoryTransform} from './src/factory_generator';
-export {FlatIndexGenerator} from './src/flat_index_generator';
 export {GeneratedShimsHostWrapper, ShimGenerator} from './src/host';
 export {SummaryGenerator} from './src/summary_generator';
