+++ packages/compiler-cli/src/ngtsc/util/src/typescript.ts
@@ -1,0 +1,18 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+const TS = /\.tsx?$/i;
+const D_TS = /\.d\.ts$/i;
+
+export function isDtsPath(filePath: string): boolean {
+  return D_TS.test(filePath);
+}
+
+export function isNonDeclarationTsPath(filePath: string): boolean {
+  return TS.test(filePath) && !D_TS.test(filePath);
+}
